ENGR00232186 ASRC TEST: modify test file for new ASRC driver
The origin ASRC test require blocks of input buffer send into ASRC
driver and the input data flow should be continuous. If not, old input
data would be send into ASRC again and again and thus noise would be
heard in output data.

In this patch, test file would send a input buffer into ASRC and wait
until the corresponding output buffer is send back to test program. In
this case, test file is not necessary promise the input data flow is
continuous.

Signed-off-by: Chen Liangjun <b36089@freescale.com>